Generacion de token y proteccion de los endpoints

+++ src/main/java/com/Procesos/negocios/AFHenr/quezB/cotrollers/UsuarioController.java
@@ -2,6 +2,7 @@ package com.Procesos.negocios.AFHenr.quezB.cotrollers;
 
 import com.Procesos.negocios.AFHenr.quezB.Services.UsuarioService;
 import com.Procesos.negocios.AFHenr.quezB.Services.UsuarioServiceImpl;
+import com.Procesos.negocios.AFHenr.quezB.Utils.JWTUtil;
 import com.Procesos.negocios.AFHenr.quezB.models.Usuario;
 import com.Procesos.negocios.AFHenr.quezB.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,9 +22,14 @@ public class UsuarioController {
 
     @Autowired
     private UsuarioService usuarioService;
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @GetMapping(value = "/usuario/{id}")
-    public ResponseEntity getUsuario(@PathVariable  Long id){
+    public ResponseEntity getUsuario(@PathVariable  Long id,@RequestHeader(value = "Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.getUserById(id);
     }
 
@@ -33,32 +39,58 @@ public class UsuarioController {
     }
 
     @GetMapping("/usuarios")
-    public ResponseEntity ListarUsuarios(){
-        return usuarioService.allUsers();
+    public ResponseEntity ListarUsuarios(@RequestHeader(value = "Authorization") String token){
+        /*if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
+        return usuarioService.allUsers();*/
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                return usuarioService.allUsers();
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
     }
 
     @GetMapping("/usuario/{nombre}/{apellidos}")
-    public ResponseEntity listarPorNombreyApellidos(@PathVariable String nombre,@PathVariable String apellidos){
+    public ResponseEntity listarPorNombreyApellidos(@PathVariable String nombre,@PathVariable String apellidos,@RequestHeader(value = "Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.allUsersByNameAndLastName(nombre,apellidos);
     }
 
     @GetMapping("/usuario/apellidos/{apellidos}")
-    public ResponseEntity listarPorApellidos(@PathVariable String apellidos){
+    public ResponseEntity listarPorApellidos(@PathVariable String apellidos,@RequestHeader(value = "Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.allUsersByLastName(apellidos);
     }
 
     @GetMapping("/usuario/nombre/{nombre}")
-    public ResponseEntity listarPorNombre(@PathVariable String nombre){
+    public ResponseEntity listarPorNombre(@PathVariable String nombre,@RequestHeader(value = "Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.allUsersByName(nombre);
     }
 
     @PutMapping("/usuario/{id}")
-    public ResponseEntity editarUsuario( @PathVariable Long id, @Valid @RequestBody  Usuario usuario){
+    public ResponseEntity editarUsuario( @PathVariable Long id, @Valid @RequestBody  Usuario usuario,@RequestHeader(value = "Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.editUser(id,usuario);
     }
 
     @DeleteMapping("/usuario/{id}")
-    public ResponseEntity eliminarUsuario(@PathVariable Long id){
+    public ResponseEntity eliminarUsuario(@PathVariable Long id,@RequestHeader(value = "Authorization") String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.deleteUserById(id);
     }
